secp256k1: Make field value add int take uint16.
This modifies the AddInt method on a field value to take a uint16
instead of a uint in order to make it impossible to misuse it since it
has a precondition of a max value of 2^26 - 1.  In practice, the code
currently only ever calls it with 7, but in order to prepare for
exporting the type, it's nicer to prevent the potential misuse.

+++ dcrec/secp256k1/field.go
@@ -527,7 +527,7 @@ func (f *fieldVal) Negate(magnitude uint32) *fieldVal {
 //
 // The field value is returned to support chaining.  This enables syntax like:
 // f.AddInt(1).Add(f2) so that f = f + 1 + f2.
-func (f *fieldVal) AddInt(ui uint) *fieldVal {
+func (f *fieldVal) AddInt(ui uint16) *fieldVal {
 	// Since the field representation intentionally provides overflow bits,
 	// it's ok to use carryless addition as the carry bit is safely part of
 	// the word and will be normalized out.
